requirements.txt
flask==1.1.2
dialogflow==1.1.0
python-telegram-bot==13.0
pandas==1.1.3
xlrd >= 1.0.0
XlsxWriter==1.3.7
